# Conflicts:
#	app/build.gradle.kts
#	app/src/main/java/com/berlin/aflami/navigation/routes/HomeRoute.kt
#	app/src/main/java/com/berlin/aflami/navigation/routes/LoginRoute.kt
#	app/src/main/java/com/berlin/aflami/navigation/routes/WatchedMediaRoute.kt
#	presentation/navigation/src/main/java/com/example/navigation/Destination.kt
#	presentation/ui/src/main/kotlin/com/berlin/aflami/navigation/NavGraph.kt
#	presentation/ui/src/main/kotlin/com/berlin/aflami/screens/authentication/LoginScreen.kt
#	presentation/ui/src/main/kotlin/com/berlin/aflami/screens/home/ContinueWatchingScreen.kt
#	presentation/ui/src/main/kotlin/com/berlin/aflami/screens/home/HomeScreen.kt
#	presentation/ui/src/main/kotlin/com/berlin/aflami/screens/mainactivity/MainActivity.kt
#	presentation/ui/src/main/kotlin/com/berlin/aflami/screens/mediadetails/screen/CastDetailsScreen.kt
#	presentation/ui/src/main/kotlin/com/berlin/aflami/screens/mediadetails/screen/MediaDetailsScreen.kt

+++ presentation/viewModel/src/main/kotlin/com/berlin/aflami/viewmodel/search/SearchInteractionListener.kt
@@ -11,7 +11,7 @@ interface SearchInteractionListener {
     fun onWorldSearchCardClicked()
     fun onActorSearchCardClicked()
     fun onTabOptionClicked(tabOption: TabOption)
-    fun onCardClicked(id: Int)
+    fun onCardClicked(id: Long)
     fun onRecentSearchClicked(query: String)
     fun onRecentSearchCleared(query: String)
     fun onAllRecentSearchesCleared()
